requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
langchain==0.0.206
openai==0.27.8
faiss-cpu==1.7.4
pypdf==3.10.0
docx2txt==0.8
unstructured==0.6.10
pinecone-client==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0